# server/pkg/api/rest/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Collecting and Aggregating server module",
        "title": "CM-Honeybee Server REST API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "latest"
    },
    "basePath": "/honeybee",
    "paths": {
...
        "/source_group/{sgId}/connection_info/{connId}/import/infra": {
            "post": {
                "description": "Import the infra information.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[Import] Import source info"
                ],
                "summary": "Import Infra",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the source group.",
                        "name": "sgId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID of the connection info.",
                        "name": "connId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully saved the infra information",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-honeybee_server_pkg_api_rest_model.SavedInfraInfo"
                        }
                    },
                    "400": {
                        "description": "Sent bad request.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-honeybee_server_pkg_api_rest_common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to save the infra information",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-honeybee_server_pkg_api_rest_common.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/source_group/{sgId}/connection_info/{connId}/import/software": {
            "post": {
                "description": "Import the software information.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[Import] Import source info"
                ],
                "summary": "Import software",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the source group.",
                        "name": "sgId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID of the connection info.",
                        "name": "connId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully saved the software information",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-honeybee_server_pkg_api_rest_model.SavedSoftwareInfo"
                        }
                    },
                    "400": {
                        "description": "Sent bad request.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-honeybee_server_pkg_api_rest_common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to save the software information",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-honeybee_server_pkg_api_rest_common.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/source_group/{sgId}/import/infra": {
            "post": {
                "description": "Import infra information for all connections in the source group.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[Import] Import source info"
                ],
                "summary": "Import Infra Source Group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the source group.",
                        "name": "sgId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully saved the infra information",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_cloud-barista_cm-honeybee_server_pkg_api_rest_model.SavedInfraInfo"
                            }
                        }
                    },
                    "400": {
                        "description": "Sent bad request.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-honeybee_server_pkg_api_rest_common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to save the infra information",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-honeybee_server_pkg_api_rest_common.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/source_group/{sgId}/import/software": {
            "post": {
                "description": "Import software information for all connections in the source group.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[Import] Import source info"
                ],
                "summary": "Import Software Source Group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the source group.",
                        "name": "sgId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully saved the software information",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_cloud-barista_cm-honeybee_server_pkg_api_rest_model.SavedInfraInfo"
                            }
                        }
                    },
                    "400": {
                        "description": "Sent bad request.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-honeybee_server_pkg_api_rest_common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to save the software information",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-honeybee_server_pkg_api_rest_common.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_cloud-barista_cm-honeybee_server_pkg_api_rest_common.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "github_com_cloud-barista_cm-honeybee_server_pkg_api_rest_model.SavedInfraInfo": {
            "type": "object",
            "required": [
                "connection_id",
                "infra_data"
            ],
            "properties": {
                "connection_id": {
                    "type": "string"
                },
                "infra_data": {
                    "type": "string"
                },
                "saved_time": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "github_com_cloud-barista_cm-honeybee_server_pkg_api_rest_model.SavedSoftwareInfo": {
            "type": "object",
            "required": [
                "connection_id",
                "software_data"
            ],
            "properties": {
                "connection_id": {
                    "type": "string"
                },
                "saved_time": {
                    "type": "string"
                },
                "software_data": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    }
}